[DSv2] Move shared parquet v2 glue to neutral package
